feature - Form implemented as an All-in-One component: - the app is fully re-writtent using this convention for the form - all tested callbacks work fine - the app seems to be much slower!!
