+++ app.py
@@ -21,6 +21,10 @@ from utils.handle_inputs import load_data, CURRENT_VERSION, DATA_DIR
 from utils.handle_inputs import availableLocations_continent, availableOptions_servers, availableOptions_country, availableOptions_region
 from utils.handle_inputs import validateInput, open_input_csv_and_comment, read_csv_input, DEFAULT_VALUES, APP_VERSION_OPTIONS_LIST
 
+from all_in_one_components.form.green_algo_form_AIO import ID_MAIN_FORM
+from all_in_one_components.form.green_algo_form_AIO_ids import GreenAlgoFormIDS
+
+form_ids = GreenAlgoFormIDS()
 
 ###################################################
 ## LOAD DATA
@@ -59,6 +63,8 @@ appVersions_options = [{'label': f'{CURRENT_VERSION} (latest)', 'value': CURRENT
 app.layout = html.Div(dash.page_container, id='fullfullPage')
 
 
+
+
 ###################################################
 # CALLBACKS #
 
@@ -71,24 +77,24 @@ app.layout = html.Div(dash.page_container, id='fullfullPage')
         ##################################################################
         ## WARNING: do not modify the order, unless modifying the order
         ## of the DEFAULT_VALUES accordingly
-        Output('runTime_hour_input','value'),
-        Output('runTime_min_input','value'),
-        Output('coreType_dropdown','value'),
-        Output('numberCPUs_input','value'),
-        Output('CPUmodel_dropdown', 'value'),
-        Output('tdpCPU_input','value'),
-        Output('numberGPUs_input','value'),
-        Output('GPUmodel_dropdown', 'value'),
-        Output('tdpGPU_input','value'),
-        Output('memory_input','value'),
-        Output('platformType_dropdown','value'),
-        Output('usageCPU_radio','value'),
-        Output('usageCPU_input','value'),
-        Output('usageGPU_radio','value'),
-        Output('usageGPU_input','value'),
-        Output('pue_radio','value'),
-        Output('PSF_radio', 'value'),
-        Output('PSF_input', 'value'),
+        Output(form_ids.runTime_hour_input(ID_MAIN_FORM),'value'),
+        Output(form_ids.runTime_min_input(ID_MAIN_FORM),'value'),
+        Output(form_ids.coreType_dropdown(ID_MAIN_FORM),'value'),
+        Output(form_ids.numberCPUs_input(ID_MAIN_FORM),'value'),
+        Output(form_ids.CPUmodel_dropdown(ID_MAIN_FORM), 'value'),
+        Output(form_ids.tdpCPU_input(ID_MAIN_FORM),'value'),
+        Output(form_ids.numberGPUs_input(ID_MAIN_FORM),'value'),
+        Output(form_ids.GPUmodel_dropdown(ID_MAIN_FORM), 'value'),
+        Output(form_ids.tdpGPU_input(ID_MAIN_FORM),'value'),
+        Output(form_ids.memory_input(ID_MAIN_FORM),'value'),
+        Output(form_ids.platformType_dropdown(ID_MAIN_FORM),'value'),
+        Output(form_ids.usageCPU_radio(ID_MAIN_FORM),'value'),
+        Output(form_ids.usageCPU_input(ID_MAIN_FORM),'value'),
+        Output(form_ids.usageGPU_radio(ID_MAIN_FORM),'value'),
+        Output(form_ids.usageGPU_input(ID_MAIN_FORM),'value'),
+        Output(form_ids.pue_radio(ID_MAIN_FORM),'value'),
+        Output(form_ids.PSF_radio(ID_MAIN_FORM), 'value'),
+        Output(form_ids.PSF_input(ID_MAIN_FORM), 'value'),
         Output('appVersions_dropdown','value'),
         Output('import-error-message', 'is_open'),
         Output('log-error-subtitle', 'children'),
@@ -100,7 +106,7 @@ app.layout = html.Div(dash.page_container, id='fullfullPage')
     ],
     [
         State('upload-data', 'filename'),
-        State('aggregate_data', 'data'),
+        State(form_ids.aggregate_data(ID_MAIN_FORM), 'data'),
     ],
     # below line is required because the 'alert' container import-error-message is not in the layout initially
     suppress_callback_exceptions=True,
@@ -180,7 +186,7 @@ def filling_from_inputs(_, upload_content, filename, current_app_state):
 @app.callback(
     Output('url_content', 'search'),
     [
-        Input('confirm_reset','submit_n_clicks'),
+        Input(form_ids.confirm_reset(ID_MAIN_FORM),'submit_n_clicks'),
     ]
 )
 def reset_url(submit_n_clicks):
@@ -221,52 +227,52 @@ def loadDataFromVersion(
 ## PLATFORM AND PROVIDER
 ########################
 
-@app.callback(
-    Output('platformType_dropdown', 'options'),
-    [Input('versioned_data','data')]
-)
-def set_platform(data):
-    '''
-    Loads platform options based on backend data.
-    '''
-    if data is not None:
-        data_dict = SimpleNamespace(**data)
-        platformType_options = [
-            {'label': k,
-             'value': v} for v, k in list(data_dict.providersTypes.items()) +
-                                     [('personalComputer', 'Personal computer')] +
-                                     [('localServer', 'Local server')]
-        ]
-        return platformType_options
-    else:
-        return []
+# @app.callback(
+#     Output('platformType_dropdown', 'options'),
+#     [Input('versioned_data','data')]
+# )
+# def set_platform(data):
+#     '''
+#     Loads platform options based on backend data.
+#     '''
+#     if data is not None:
+#         data_dict = SimpleNamespace(**data)
+#         platformType_options = [
+#             {'label': k,
+#              'value': v} for v, k in list(data_dict.providersTypes.items()) +
+#                                      [('personalComputer', 'Personal computer')] +
+#                                      [('localServer', 'Local server')]
+#         ]
+#         return platformType_options
+#     else:
+#         return []
+
+# @app.callback(
+#     Output('provider_dropdown_div', 'style'),
+#     [Input('platformType_dropdown', 'value')]
+# )
+# def set_providers(selected_platform):
+#     '''
+#     Shows or hide the "providers" box, based on the platform selected.
+#     '''
+#     if selected_platform in ['cloudComputing']:
+#         # Only Cloud Computing need the providers box
+#         outputStyle = {'display': 'block'}
+#     else:
+#         outputStyle = {'display': 'none'}
+
+#     return outputStyle
 
 @app.callback(
-    Output('provider_dropdown_div', 'style'),
-    [Input('platformType_dropdown', 'value')]
-)
-def set_providers(selected_platform):
-    '''
-    Shows or hide the "providers" box, based on the platform selected.
-    '''
-    if selected_platform in ['cloudComputing']:
-        # Only Cloud Computing need the providers box
-        outputStyle = {'display': 'block'}
-    else:
-        outputStyle = {'display': 'none'}
-
-    return outputStyle
-
-@app.callback(
-    Output('provider_dropdown','value'),
+    Output(form_ids.provider_dropdown(ID_MAIN_FORM),'value'),
     [
-        Input('platformType_dropdown', 'value'),
+        Input(form_ids.platformType_dropdown(ID_MAIN_FORM), 'value'),
         Input('versioned_data','data'),
         Input('upload-data', 'contents'),
     ],
     [
         State('upload-data', 'filename'),
-        State('provider_dropdown', 'value'),
+        State(form_ids.provider_dropdown(ID_MAIN_FORM), 'value'),
     ]
 )
 def set_provider(_, versioned_data, upload_content, filename, prev_provider):
@@ -288,194 +294,194 @@ def set_provider(_, versioned_data, upload_content, filename, prev_provider):
                 
     return 'gcp'     
 
-@app.callback(
-    Output('provider_dropdown', 'options'),
-    [
-        Input('platformType_dropdown', 'value'),
-        Input('versioned_data','data')
-    ],
-)
-def set_providers(selected_platform, data):
-    '''
-    List options for the "provider" box.
-    '''
-    if data is not None:
-        data_dict = SimpleNamespace(**data)
+# @app.callback(
+#     Output('provider_dropdown', 'options'),
+#     [
+#         Input('platformType_dropdown', 'value'),
+#         Input('versioned_data','data')
+#     ],
+# )
+# def set_providers(selected_platform, data):
+#     '''
+#     List options for the "provider" box.
+#     '''
+#     if data is not None:
+#         data_dict = SimpleNamespace(**data)
 
-        providers_dict = data_dict.platformName_byType.get(selected_platform)
-        if providers_dict is not None:
-            availableOptions = list(providers_dict.items())
-        else:
-            availableOptions = []
+#         providers_dict = data_dict.platformName_byType.get(selected_platform)
+#         if providers_dict is not None:
+#             availableOptions = list(providers_dict.items())
+#         else:
+#             availableOptions = []
 
-        listOptions = [
-            {'label': v, 'value': k} for k, v in availableOptions + [("other","Other")]
-        ]
-        return listOptions
-    else:
-        return []
+#         listOptions = [
+#             {'label': v, 'value': k} for k, v in availableOptions + [("other","Other")]
+#         ]
+#         return listOptions
+#     else:
+#         return []
 
 
 ## COMPUTING CORES
 ##################
 
-@app.callback(
-    Output('coreType_dropdown', 'options'),
-    [
-        Input('provider_dropdown', 'value'),
-        Input('platformType_dropdown', 'value'),
-        Input('versioned_data','data')
-    ])
-def set_coreType_options(_, __, data):
-    '''
-    List of options for coreType (CPU or GPU), based on the platform/provider selected.
-    Not really useful so far because we have no specific core types for a given provider.
-    '''
-    if data is not None:
-        data_dict = SimpleNamespace(**data)
+# @app.callback(
+#     Output('coreType_dropdown', 'options'),
+#     [
+#         Input('provider_dropdown', 'value'),
+#         Input('platformType_dropdown', 'value'),
+#         Input('versioned_data','data')
+#     ])
+# def set_coreType_options(_, __, data):
+#     '''
+#     List of options for coreType (CPU or GPU), based on the platform/provider selected.
+#     Not really useful so far because we have no specific core types for a given provider.
+#     '''
+#     if data is not None:
+#         data_dict = SimpleNamespace(**data)
 
-        availableOptions = data_dict.cores_dict.keys()
-        listOptions = [{'label': k, 'value': k} for k in list(sorted(availableOptions))+['Both']]
-        return listOptions
-    else:
-        return []
+#         availableOptions = data_dict.cores_dict.keys()
+#         listOptions = [{'label': k, 'value': k} for k in list(sorted(availableOptions))+['Both']]
+#         return listOptions
+#     else:
+#         return []
 
-@app.callback(
-    [
-        Output('CPUmodel_dropdown', 'options'),
-        Output('GPUmodel_dropdown', 'options')
-    ],
-    [Input('versioned_data','data')]
-)
-def set_coreOptions(data):
-    '''
-    List of options for core models.
-    '''
-    if data is not None:
-        data_dict = SimpleNamespace(**data)
+# @app.callback(
+#     [
+#         Output('CPUmodel_dropdown', 'options'),
+#         Output('GPUmodel_dropdown', 'options')
+#     ],
+#     [Input('versioned_data','data')]
+# )
+# def set_coreOptions(data):
+#     '''
+#     List of options for core models.
+#     '''
+#     if data is not None:
+#         data_dict = SimpleNamespace(**data)
 
-        coreModels_options = dict()
-        for coreType in ['CPU', 'GPU']:
-            availableOptions = sorted(list(data_dict.cores_dict[coreType].keys()))
-            availableOptions = put_value_first(availableOptions, 'Any')
-            coreModels_options[coreType] = [
-                {'label': k, 'value': v} for k, v in list(zip(availableOptions, availableOptions)) +
-                [("Other", "other")]
-            ]
+#         coreModels_options = dict()
+#         for coreType in ['CPU', 'GPU']:
+#             availableOptions = sorted(list(data_dict.cores_dict[coreType].keys()))
+#             availableOptions = put_value_first(availableOptions, 'Any')
+#             coreModels_options[coreType] = [
+#                 {'label': k, 'value': v} for k, v in list(zip(availableOptions, availableOptions)) +
+#                 [("Other", "other")]
+#             ]
 
-        return coreModels_options['CPU'], coreModels_options['GPU']
+#         return coreModels_options['CPU'], coreModels_options['GPU']
 
-    else:
-        return [],[]
+#     else:
+#         return [],[]
 
-@app.callback(
-    [
-        Output('CPU_div', 'style'),
-        Output('title_CPU', 'style'),
-        Output('usageCPU_div', 'style'),
-        Output('GPU_div', 'style'),
-        Output('title_GPU', 'style'),
-        Output('usageGPU_div', 'style'),
-    ],
-    [
-        Input('coreType_dropdown', 'value')
-    ]
-)
-def show_CPUGPUdiv(selected_coreType):
-    '''
-    Shows or hides the CPU/GPU input blocks (and the titles) based on the selected core type.
-    '''
-    show = {'display': 'block'}
-    showFlex = {'display': 'flex'}
-    hide = {'display': 'none'}
-    if selected_coreType == 'CPU':
-        return show, hide, showFlex, hide, hide, hide
-    elif selected_coreType == 'GPU':
-        return hide, hide, hide, show, hide, showFlex
-    else:
-        return show, show, showFlex, show, show, showFlex
+# @app.callback(
+#     [
+#         Output('CPU_div', 'style'),
+#         Output('title_CPU', 'style'),
+#         Output('usageCPU_div', 'style'),
+#         Output('GPU_div', 'style'),
+#         Output('title_GPU', 'style'),
+#         Output('usageGPU_div', 'style'),
+#     ],
+#     [
+#         Input('coreType_dropdown', 'value')
+#     ]
+# )
+# def show_CPUGPUdiv(selected_coreType):
+#     '''
+#     Shows or hides the CPU/GPU input blocks (and the titles) based on the selected core type.
+#     '''
+#     show = {'display': 'block'}
+#     showFlex = {'display': 'flex'}
+#     hide = {'display': 'none'}
+#     if selected_coreType == 'CPU':
+#         return show, hide, showFlex, hide, hide, hide
+#     elif selected_coreType == 'GPU':
+#         return hide, hide, hide, show, hide, showFlex
+#     else:
+#         return show, show, showFlex, show, show, showFlex
 
-@app.callback(
-    Output('tdpCPU_div', 'style'),
-    [
-        Input('CPUmodel_dropdown', 'value'),
-    ]
-)
-def display_TDP4CPU(selected_coreModel):
-    '''
-    Shows or hides the CPU TDP input box.
-    '''
-    if selected_coreModel == "other":
-        return {'display': 'flex'}
-    else:
-        return {'display': 'none'}
+# @app.callback(
+#     Output('tdpCPU_div', 'style'),
+#     [
+#         Input('CPUmodel_dropdown', 'value'),
+#     ]
+# )
+# def display_TDP4CPU(selected_coreModel):
+#     '''
+#     Shows or hides the CPU TDP input box.
+#     '''
+#     if selected_coreModel == "other":
+#         return {'display': 'flex'}
+#     else:
+#         return {'display': 'none'}
 
-@app.callback(
-    Output('tdpGPU_div', 'style'),
-    [
-        Input('GPUmodel_dropdown', 'value'),
-    ]
-)
-def display_TDP4GPU(selected_coreModel):
-    '''
-    Shows or hides the GPU TDP input box.
-    '''
-    if selected_coreModel == "other":
-        return {'display': 'flex'}
-    else:
-        return {'display': 'none'}
+# @app.callback(
+#     Output('tdpGPU_div', 'style'),
+#     [
+#         Input('GPUmodel_dropdown', 'value'),
+#     ]
+# )
+# def display_TDP4GPU(selected_coreModel):
+#     '''
+#     Shows or hides the GPU TDP input box.
+#     '''
+#     if selected_coreModel == "other":
+#         return {'display': 'flex'}
+#     else:
+#         return {'display': 'none'}
 
 
 ## LOCATION AND SERVER
 ######################
 
-@app.callback(
-    [
-        Output('location_div', 'style'),
-        Output('server_div', 'style'),
-    ],
-    [
-        Input('platformType_dropdown', 'value'),
-        Input('provider_dropdown', 'value'),
-        Input('server_dropdown','value'),
-        Input('versioned_data','data')
-    ]
-)
-def display_location(selected_platform, selected_provider, selected_server, data):
-    '''
-    Shows either LOCATION or SERVER depending on the platform.
-    '''
-    if data is not None:
-        data_dict = SimpleNamespace(**data)
-        providers_withoutDC = data_dict.providers_withoutDC
-    else:
-        providers_withoutDC = []
+# @app.callback(
+#     [
+#         Output('location_div', 'style'),
+#         Output('server_div', 'style'),
+#     ],
+#     [
+#         Input('platformType_dropdown', 'value'),
+#         Input('provider_dropdown', 'value'),
+#         Input('server_dropdown','value'),
+#         Input('versioned_data','data')
+#     ]
+# )
+# def display_location(selected_platform, selected_provider, selected_server, data):
+#     '''
+#     Shows either LOCATION or SERVER depending on the platform.
+#     '''
+#     if data is not None:
+#         data_dict = SimpleNamespace(**data)
+#         providers_withoutDC = data_dict.providers_withoutDC
+#     else:
+#         providers_withoutDC = []
 
-    show = {'display': 'flex'}
-    hide = {'display': 'none'}
-    if selected_platform == 'cloudComputing':
-        if selected_provider in ['other'] + providers_withoutDC:
-            return show, hide
-        elif selected_server == 'other':
-            return show, show
-        else:
-            return hide, show
-    else:
-        return show, hide
+#     show = {'display': 'flex'}
+#     hide = {'display': 'none'}
+#     if selected_platform == 'cloudComputing':
+#         if selected_provider in ['other'] + providers_withoutDC:
+#             return show, hide
+#         elif selected_server == 'other':
+#             return show, show
+#         else:
+#             return hide, show
+#     else:
+#         return show, hide
     
 
-### SERVER (only for Cloud computing for now)
+# ### SERVER (only for Cloud computing for now)
 
 @app.callback(
-    Output('server_continent_dropdown','value'),
+    Output(form_ids.server_continent_dropdown(ID_MAIN_FORM),'value'),
     [
-        Input('provider_dropdown', 'value'),
+        Input(form_ids.provider_dropdown(ID_MAIN_FORM), 'value'),
         Input('versioned_data','data'),
         Input('upload-data', 'contents'),
     ],
     [
         State('upload-data', 'filename'),
-        State('server_continent_dropdown', 'value'),
+        State(form_ids.server_continent_dropdown(ID_MAIN_FORM), 'value'),
     ]
 )
 def set_serverContinents_value(selected_provider, versioned_data, upload_content, filename, prev_server_continent):
@@ -515,48 +521,48 @@ def set_serverContinents_value(selected_provider, versioned_data, upload_content
             defaultValue = None
     return defaultValue
 
-@app.callback(
-    Output('server_continent_dropdown','options'),
-    [
-        Input('provider_dropdown', 'value'),
-        Input('versioned_data','data')
-    ]
-)
-def set_serverContinents_options(selected_provider, data):
-    '''
-    List of options and default value for server's continent, based on the provider
-    '''
-    availableOptions = availableLocations_continent(selected_provider, data=data)
-    listOptions = [{'label': k, 'value': k} for k in sorted(availableOptions)] + [{'label': 'Other', 'value': 'other'}]
-    return listOptions
+# @app.callback(
+#     Output('server_continent_dropdown','options'),
+#     [
+#         Input('provider_dropdown', 'value'),
+#         Input('versioned_data','data')
+#     ]
+# )
+# def set_serverContinents_options(selected_provider, data):
+#     '''
+#     List of options and default value for server's continent, based on the provider
+#     '''
+#     availableOptions = availableLocations_continent(selected_provider, data=data)
+#     listOptions = [{'label': k, 'value': k} for k in sorted(availableOptions)] + [{'label': 'Other', 'value': 'other'}]
+#     return listOptions
+
+# @app.callback(
+#     Output('server_dropdown','style'),
+#     [
+#         Input('server_continent_dropdown', 'value')
+#     ]
+# )
+# def set_server_style(selected_continent):
+#     '''
+#     Show or not the choice of servers, don't if continent is on "Other"
+#     '''
+#     if selected_continent == 'other':
+#         return {'display': 'none'}
+
+#     else:
+#         return {'display': 'block'}
 
 @app.callback(
-    Output('server_dropdown','style'),
+    Output(form_ids.server_dropdown(ID_MAIN_FORM),'value'),
     [
-        Input('server_continent_dropdown', 'value')
-    ]
-)
-def set_server_style(selected_continent):
-    '''
-    Show or not the choice of servers, don't if continent is on "Other"
-    '''
-    if selected_continent == 'other':
-        return {'display': 'none'}
-
-    else:
-        return {'display': 'block'}
-
-@app.callback(
-    Output('server_dropdown','value'),
-    [
-        Input('provider_dropdown', 'value'),
-        Input('server_continent_dropdown', 'value'),
+        Input(form_ids.provider_dropdown(ID_MAIN_FORM), 'value'),
+        Input(form_ids.server_continent_dropdown(ID_MAIN_FORM), 'value'),
         Input('versioned_data','data'),
         Input('upload-data', 'contents'),
     ],
     [
         State('upload-data', 'filename'),
-        State('server_dropdown','value'),
+        State(form_ids.server_dropdown(ID_MAIN_FORM),'value'),
     ]
 )
 def set_server_value(selected_provider, selected_continent, versioned_data, upload_content, filename, prev_server_value):
@@ -600,51 +606,51 @@ def set_server_value(selected_provider, selected_continent, versioned_data, uplo
         defaultValue = None
     return defaultValue
 
+# @app.callback(
+#     Output('server_dropdown','options'),
+#     [
+#         Input('provider_dropdown', 'value'),
+#         Input('server_continent_dropdown', 'value'),
+#         Input('versioned_data','data')
+#     ]
+# )
+# def set_server_options(selected_provider,selected_continent, data):
+#     '''
+#     List of options for servers, based on provider and continent
+#     '''
+#     availableOptions = availableOptions_servers(selected_provider,selected_continent,data=data)
+#     listOptions = [{'label': k['Name'], 'value': k['name_unique']} for k in availableOptions + [{'Name':"other", 'name_unique':'other'}]]
+
+#     return listOptions
+
+
+# ## LOCATION (only for local server, personal device or "other" cloud server)
+
+# @app.callback(
+#     Output('location_continent_dropdown', 'options'),
+#     [Input('versioned_data','data')]
+# )
+# def set_continentOptions(data):
+#     if data is not None:
+#         data_dict = SimpleNamespace(**data)
+
+#         continentsList = list(data_dict.CI_dict_byName.keys())
+#         continentsDict = [{'label': k, 'value': k} for k in sorted(continentsList)]
+
+#         return continentsDict
+#     else:
+#         return []
+
 @app.callback(
-    Output('server_dropdown','options'),
+    Output(form_ids.location_continent_dropdown(ID_MAIN_FORM), 'value'),
     [
-        Input('provider_dropdown', 'value'),
-        Input('server_continent_dropdown', 'value'),
-        Input('versioned_data','data')
-    ]
-)
-def set_server_options(selected_provider,selected_continent, data):
-    '''
-    List of options for servers, based on provider and continent
-    '''
-    availableOptions = availableOptions_servers(selected_provider,selected_continent,data=data)
-    listOptions = [{'label': k['Name'], 'value': k['name_unique']} for k in availableOptions + [{'Name':"other", 'name_unique':'other'}]]
-
-    return listOptions
-
-
-## LOCATION (only for local server, personal device or "other" cloud server)
-
-@app.callback(
-    Output('location_continent_dropdown', 'options'),
-    [Input('versioned_data','data')]
-)
-def set_continentOptions(data):
-    if data is not None:
-        data_dict = SimpleNamespace(**data)
-
-        continentsList = list(data_dict.CI_dict_byName.keys())
-        continentsDict = [{'label': k, 'value': k} for k in sorted(continentsList)]
-
-        return continentsDict
-    else:
-        return []
-
-@app.callback(
-    Output('location_continent_dropdown', 'value'),
-    [
-        Input('server_continent_dropdown','value'),
-        Input('server_div', 'style'),
+        Input(form_ids.server_continent_dropdown(ID_MAIN_FORM),'value'),
+        Input(form_ids.server_div(ID_MAIN_FORM), 'style'),
         Input('upload-data', 'contents'),
     ],
     [
         State('upload-data', 'filename'),
-        State('location_continent_dropdown', 'value'),
+        State(form_ids.location_continent_dropdown(ID_MAIN_FORM), 'value'),
         State('versioned_data','data'),
     ]
 )
@@ -680,18 +686,18 @@ def set_continent_value(selected_serverContinent, display_server, upload_content
 
 @app.callback(
     [
-        Output('location_country_dropdown', 'options'),
-        Output('location_country_dropdown', 'value'),
-        Output('location_country_dropdown_div', 'style'),
+        Output(form_ids.location_country_dropdown(ID_MAIN_FORM), 'options'),
+        Output(form_ids.location_country_dropdown(ID_MAIN_FORM), 'value'),
+        Output(form_ids.location_country_dropdown_div(ID_MAIN_FORM), 'style'),
     ],
     [
-        Input('location_continent_dropdown', 'value'),
+        Input(form_ids.location_continent_dropdown(ID_MAIN_FORM), 'value'),
         Input('versioned_data','data'),
         Input('upload-data', 'contents'),
     ],
     [
         State('upload-data', 'filename'),
-        State('location_country_dropdown', 'value')
+        State(form_ids.location_country_dropdown(ID_MAIN_FORM), 'value')
     ]
 )
 def set_countries_options(selected_continent, versioned_data, upload_content, filename, prev_selectedCountry):
@@ -742,19 +748,19 @@ def set_countries_options(selected_continent, versioned_data, upload_content, fi
 
 @app.callback(
     [
-        Output('location_region_dropdown', 'options'),
-        Output('location_region_dropdown', 'value'),
-        Output('location_region_dropdown_div', 'style'),
+        Output(form_ids.location_region_dropdown(ID_MAIN_FORM), 'options'),
+        Output(form_ids.location_region_dropdown(ID_MAIN_FORM), 'value'),
+        Output(form_ids.location_region_dropdown_div(ID_MAIN_FORM), 'style'),
     ],
     [
-        Input('location_continent_dropdown', 'value'),
-        Input('location_country_dropdown', 'value'),
+        Input(form_ids.location_continent_dropdown(ID_MAIN_FORM), 'value'),
+        Input(form_ids.location_country_dropdown(ID_MAIN_FORM), 'value'),
         Input('versioned_data','data'),
         Input('upload-data', 'contents'),
     ],
     [
         State('upload-data', 'filename'),
-        State('location_region_dropdown', 'value'),
+        State(form_ids.location_region_dropdown(ID_MAIN_FORM), 'value'),
     ]
 
 )
@@ -808,106 +814,106 @@ def set_regions_options(selected_continent, selected_country, versioned_data, up
     return listOptions, defaultValue, region_style
 
 
-## USAGE FACTORS
-################
+# ## USAGE FACTORS
+# ################
+
+# @app.callback(
+#     Output('usageCPU_input','style'),
+#     [
+#         Input('usageCPU_radio', 'value'),
+#         Input('usageCPU_input', 'disabled')
+#     ]
+# )
+# def display_usage_input(answer_usage, disabled):
+#     '''
+#     Show or hide the usage factor input box, based on Yes/No input
+#     '''
+#     if answer_usage == 'No':
+#         out = {'display': 'none'}
+#     else:
+#         out = {'display': 'block'}
+
+#     if disabled:
+#         out['background-color'] = MY_COLORS['boxesColor']
+
+#     return out
+
+# @app.callback(
+#     Output('usageGPU_input','style'),
+#     [
+#         Input('usageGPU_radio', 'value'),
+#         Input('usageGPU_input', 'disabled')
+#     ]
+# )
+# def display_usage_input(answer_usage, disabled):
+#     '''
+#     Show or hide the usage factor input box, based on Yes/No input
+#     '''
+#     if answer_usage == 'No':
+#         out = {'display': 'none'}
+#     else:
+#         out = {'display': 'block'}
+
+#     if disabled:
+#         out['background-color'] = MY_COLORS['boxesColor']
+
+#     return out
+
+
+# ## PUE INPUTS
+# ################
+
+# @app.callback(
+#     Output('PUEquestion_div','style'),
+#     [
+#         Input('location_region_dropdown','value'),
+#         Input('platformType_dropdown', 'value'),
+#         Input('provider_dropdown', 'value'),
+#         Input('server_dropdown', 'value')
+#     ]
+# )
+# def display_pue_question(_, selected_platform, selected_provider, selected_server):
+#     '''
+#     Shows or hides the PUE question depending on the platform
+#     '''
+#     if selected_platform == 'localServer':
+#         return {'display': 'flex'}
+#     elif (selected_platform == 'cloudComputing')&((selected_provider == 'other')|(selected_server == 'other')):
+#         return {'display': 'flex'}
+#     else:
+#         return {'display': 'none'}
+
+# @app.callback(
+#     Output('PUE_input','style'),
+#     [
+#         Input('pue_radio', 'value'),
+#         Input('PUE_input','disabled')
+#     ]
+# )
+# def display_pue_input(answer_pue, disabled):
+#     '''
+#     Shows or hides the PUE input box
+#     '''
+#     if answer_pue == 'No':
+#         out = {'display': 'none'}
+#     else:
+#         out = {'display': 'block'}
+
+#     if disabled:
+#         out['background-color'] = MY_COLORS['boxesColor']
+
+#     return out
 
 @app.callback(
-    Output('usageCPU_input','style'),
+    Output(form_ids.PUE_input(ID_MAIN_FORM),'value'),
     [
-        Input('usageCPU_radio', 'value'),
-        Input('usageCPU_input', 'disabled')
-    ]
-)
-def display_usage_input(answer_usage, disabled):
-    '''
-    Show or hide the usage factor input box, based on Yes/No input
-    '''
-    if answer_usage == 'No':
-        out = {'display': 'none'}
-    else:
-        out = {'display': 'block'}
-
-    if disabled:
-        out['background-color'] = MY_COLORS['boxesColor']
-
-    return out
-
-@app.callback(
-    Output('usageGPU_input','style'),
-    [
-        Input('usageGPU_radio', 'value'),
-        Input('usageGPU_input', 'disabled')
-    ]
-)
-def display_usage_input(answer_usage, disabled):
-    '''
-    Show or hide the usage factor input box, based on Yes/No input
-    '''
-    if answer_usage == 'No':
-        out = {'display': 'none'}
-    else:
-        out = {'display': 'block'}
-
-    if disabled:
-        out['background-color'] = MY_COLORS['boxesColor']
-
-    return out
-
-
-## PUE INPUTS
-################
-
-@app.callback(
-    Output('PUEquestion_div','style'),
-    [
-        Input('location_region_dropdown','value'),
-        Input('platformType_dropdown', 'value'),
-        Input('provider_dropdown', 'value'),
-        Input('server_dropdown', 'value')
-    ]
-)
-def display_pue_question(_, selected_platform, selected_provider, selected_server):
-    '''
-    Shows or hides the PUE question depending on the platform
-    '''
-    if selected_platform == 'localServer':
-        return {'display': 'flex'}
-    elif (selected_platform == 'cloudComputing')&((selected_provider == 'other')|(selected_server == 'other')):
-        return {'display': 'flex'}
-    else:
-        return {'display': 'none'}
-
-@app.callback(
-    Output('PUE_input','style'),
-    [
-        Input('pue_radio', 'value'),
-        Input('PUE_input','disabled')
-    ]
-)
-def display_pue_input(answer_pue, disabled):
-    '''
-    Shows or hides the PUE input box
-    '''
-    if answer_pue == 'No':
-        out = {'display': 'none'}
-    else:
-        out = {'display': 'block'}
-
-    if disabled:
-        out['background-color'] = MY_COLORS['boxesColor']
-
-    return out
-
-@app.callback(
-    Output('PUE_input','value'),
-    [
-        Input('pue_radio', 'value'),
+        Input(form_ids.pue_radio(ID_MAIN_FORM), 'value'),
         Input('versioned_data','data'),
         Input('upload-data', 'contents'),
     ],
     [
         State('upload-data', 'filename'),
-        State('PUE_input','value'),
+        State(form_ids.PUE_input(ID_MAIN_FORM),'value'),
     ]
 )
 def set_PUE(radio, versioned_data, upload_content, filename, prev_pue):
@@ -937,38 +943,38 @@ def set_PUE(radio, versioned_data, upload_content, filename, prev_pue):
     return defaultPUE
 
 
-## PSF INPUTS
-##############
+# ## PSF INPUTS
+# ##############
+
+# @app.callback(
+#     Output('PSF_input','style'),
+#     [
+#         Input('PSF_radio', 'value'),
+#         Input('PSF_input', 'disabled')
+#     ]
+# )
+# def display_PSF_input(answer_PSF, disabled):
+#     '''
+#     Shows or hides the PSF input box
+#     '''
+#     if answer_PSF == 'No':
+#         out = {'display': 'none'}
+#     else:
+#         out = {'display': 'block'}
+
+#     if disabled:
+#         out['background-color'] = MY_COLORS['boxesColor']
+
+#     return out
+
+
+# ## RESET AND APP VERSIONS
+# #########################
 
 @app.callback(
-    Output('PSF_input','style'),
+    Output(form_ids.confirm_reset(ID_MAIN_FORM),'displayed'),
     [
-        Input('PSF_radio', 'value'),
-        Input('PSF_input', 'disabled')
-    ]
-)
-def display_PSF_input(answer_PSF, disabled):
-    '''
-    Shows or hides the PSF input box
-    '''
-    if answer_PSF == 'No':
-        out = {'display': 'none'}
-    else:
-        out = {'display': 'block'}
-
-    if disabled:
-        out['background-color'] = MY_COLORS['boxesColor']
-
-    return out
-
-
-## RESET AND APP VERSIONS
-#########################
-
-@app.callback(
-    Output('confirm_reset','displayed'),
-    [
-        Input('reset_link','n_clicks')
+        Input(form_ids.reset_link(ID_MAIN_FORM),'n_clicks')
     ]
 )
 def display_confirm(clicks):
@@ -999,354 +1005,354 @@ def display_oldVersion(clicks, version, oldStyle):
         return oldStyle
 
 
-## PROCESS INPUTS
-#################
+# ## PROCESS INPUTS
+# #################
 
-@app.callback(
-    Output("aggregate_data", "data"),
-    [
-        Input('versioned_data','data'),
-        Input("coreType_dropdown", "value"),
-        Input("numberCPUs_input", "value"),
-        Input("CPUmodel_dropdown", "value"),
-        Input("tdpCPU_div", "style"),
-        Input("tdpCPU_input", "value"),
-        Input("numberGPUs_input", "value"),
-        Input("GPUmodel_dropdown", "value"),
-        Input("tdpGPU_div", "style"),
-        Input("tdpGPU_input", "value"),
-        Input("memory_input", "value"),
-        Input("runTime_hour_input", "value"),
-        Input("runTime_min_input", "value"),
-        Input("location_continent_dropdown", "value"),
-        Input("location_country_dropdown", "value"),
-        Input("location_region_dropdown", "value"),
-        Input("server_continent_dropdown", "value"),
-        Input("server_dropdown", "value"),
-        Input('location_div', 'style'),
-        Input('server_div','style'),
-        Input("usageCPU_radio", "value"),
-        Input("usageCPU_input", "value"),
-        Input("usageGPU_radio", "value"),
-        Input("usageGPU_input", "value"),
-        Input('PUEquestion_div','style'),
-        Input("pue_radio", "value"),
-        Input("PUE_input", "value"),
-        Input("PSF_radio", "value"),
-        Input("PSF_input", "value"),
-        Input('platformType_dropdown', 'value'),
-        Input('provider_dropdown', 'value'),
-        Input('provider_dropdown_div', 'style'),
-    ],
-)
-def aggregate_input_values(data, coreType, n_CPUcores, CPUmodel, tdpCPUstyle, tdpCPU, n_GPUs, GPUmodel, tdpGPUstyle, tdpGPU,
-                           memory, runTime_hours, runTime_min, locationContinent, locationCountry, locationRegion,
-                           serverContinent, server, locationStyle, serverStyle, usageCPUradio, usageCPU, usageGPUradio, usageGPU,
-                           PUEdivStyle, PUEradio, PUE, PSFradio, PSF, selected_platform, selected_provider, providerStyle):
-    '''
-    Computes all the metrics and gathers the information provided by the inputs of the form.
-    '''
-    output = dict()
+# @app.callback(
+#     Output("aggregate_data", "data"),
+#     [
+#         Input('versioned_data','data'),
+#         Input("coreType_dropdown", "value"),
+#         Input("numberCPUs_input", "value"),
+#         Input("CPUmodel_dropdown", "value"),
+#         Input("tdpCPU_div", "style"),
+#         Input("tdpCPU_input", "value"),
+#         Input("numberGPUs_input", "value"),
+#         Input("GPUmodel_dropdown", "value"),
+#         Input("tdpGPU_div", "style"),
+#         Input("tdpGPU_input", "value"),
+#         Input("memory_input", "value"),
+#         Input("runTime_hour_input", "value"),
+#         Input("runTime_min_input", "value"),
+#         Input("location_continent_dropdown", "value"),
+#         Input("location_country_dropdown", "value"),
+#         Input("location_region_dropdown", "value"),
+#         Input("server_continent_dropdown", "value"),
+#         Input("server_dropdown", "value"),
+#         Input('location_div', 'style'),
+#         Input('server_div','style'),
+#         Input("usageCPU_radio", "value"),
+#         Input("usageCPU_input", "value"),
+#         Input("usageGPU_radio", "value"),
+#         Input("usageGPU_input", "value"),
+#         Input('PUEquestion_div','style'),
+#         Input("pue_radio", "value"),
+#         Input("PUE_input", "value"),
+#         Input("PSF_radio", "value"),
+#         Input("PSF_input", "value"),
+#         Input('platformType_dropdown', 'value'),
+#         Input('provider_dropdown', 'value'),
+#         Input('provider_dropdown_div', 'style'),
+#     ],
+# )
+# def aggregate_input_values(data, coreType, n_CPUcores, CPUmodel, tdpCPUstyle, tdpCPU, n_GPUs, GPUmodel, tdpGPUstyle, tdpGPU,
+#                            memory, runTime_hours, runTime_min, locationContinent, locationCountry, locationRegion,
+#                            serverContinent, server, locationStyle, serverStyle, usageCPUradio, usageCPU, usageGPUradio, usageGPU,
+#                            PUEdivStyle, PUEradio, PUE, PSFradio, PSF, selected_platform, selected_provider, providerStyle):
+#     '''
+#     Computes all the metrics and gathers the information provided by the inputs of the form.
+#     '''
+#     output = dict()
 
-    #############################################
-    ### PREPROCESS: check if computations can be performed
+#     #############################################
+#     ### PREPROCESS: check if computations can be performed
 
-    notReady = False
+#     notReady = False
 
-    ### Runtime
-    test_runTime = 0
-    if runTime_hours is None:
-        actual_runTime_hours = 0
-        test_runTime += 1
-    else:
-        actual_runTime_hours = runTime_hours
+#     ### Runtime
+#     test_runTime = 0
+#     if runTime_hours is None:
+#         actual_runTime_hours = 0
+#         test_runTime += 1
+#     else:
+#         actual_runTime_hours = runTime_hours
 
-    if runTime_min is None:
-        actual_runTime_min = 0
-        test_runTime += 1
-    else:
-        actual_runTime_min = runTime_min
-    runTime = actual_runTime_hours + actual_runTime_min/60.
+#     if runTime_min is None:
+#         actual_runTime_min = 0
+#         test_runTime += 1
+#     else:
+#         actual_runTime_min = runTime_min
+#     runTime = actual_runTime_hours + actual_runTime_min/60.
 
-    ### Core type
-    if coreType is None:
-        notReady = True
-    elif (coreType in ['CPU','Both'])&((n_CPUcores is None)|(CPUmodel is None)):
-        notReady = True
-    elif (coreType in ['GPU','Both'])&((n_GPUs is None)|(GPUmodel is None)):
-        notReady = True
+#     ### Core type
+#     if coreType is None:
+#         notReady = True
+#     elif (coreType in ['CPU','Both'])&((n_CPUcores is None)|(CPUmodel is None)):
+#         notReady = True
+#     elif (coreType in ['GPU','Both'])&((n_GPUs is None)|(GPUmodel is None)):
+#         notReady = True
 
-    ### Versioned data
-    if data is not None:
-        data_dict = SimpleNamespace(**data)
-        version = data_dict.version
-    else:
-        notReady = True
+#     ### Versioned data
+#     if data is not None:
+#         data_dict = SimpleNamespace(**data)
+#         version = data_dict.version
+#     else:
+#         notReady = True
 
-    ### Location
-    if is_shown(locationStyle):
-        # this means the "location" input is shown, so we use location instead of server
-        locationVar = locationRegion
-    elif (server is None) | (server == 'other') | (data is None):
-        locationVar = None
-    else:
-        locationVar = data_dict.datacenters_dict_byName[server]['location']
+#     ### Location
+#     if is_shown(locationStyle):
+#         # this means the "location" input is shown, so we use location instead of server
+#         locationVar = locationRegion
+#     elif (server is None) | (server == 'other') | (data is None):
+#         locationVar = None
+#     else:
+#         locationVar = data_dict.datacenters_dict_byName[server]['location']
 
-    ### Platform
-    if selected_platform is None:
-        notReady = True
-    elif (selected_platform == 'cloudComputing')&(selected_provider is None):
-        notReady = True
+#     ### Platform
+#     if selected_platform is None:
+#         notReady = True
+#     elif (selected_platform == 'cloudComputing')&(selected_provider is None):
+#         notReady = True
 
-    ### Other required inputs
-    if (memory is None)|(tdpCPU is None)|(tdpGPU is None)|(locationVar is None)| \
-            (usageCPU is None)|(usageGPU is None)|(PUE is None)|(PSF is None):
-        notReady = True
+#     ### Other required inputs
+#     if (memory is None)|(tdpCPU is None)|(tdpGPU is None)|(locationVar is None)| \
+#             (usageCPU is None)|(usageGPU is None)|(PUE is None)|(PSF is None):
+#         notReady = True
 
-    ### If any of the required inputs is note ready: do not compute
-    if notReady:
-        output['coreType'] = None
-        output['CPUmodel'] = None
-        output['numberCPUs'] = None
-        output['usageCPU'] = None
-        output['usageCPUradio'] = None
-        output['tdpCPU'] = None
-        output['GPUmodel'] = None
-        output['numberGPUs'] = None
-        output['tdpGPU'] = None
-        output['usageGPU'] = None
-        output['usageGPUradio'] = None
-        output['GPUpower'] = None
-        output['memory'] = None
-        output['runTime_hour'] = None
-        output['runTime_min'] = None
-        output['runTime'] = None
-        output['platformType'] = None
-        output['location'] = None
-        output['carbonIntensity'] = None
-        output['PUE'] = None
-        output['PUEradio'] = None
-        output['PSF'] = None
-        output['PSFradio'] = None
-        output['carbonEmissions'] = 0
-        output['CE_CPU'] = 0
-        output['CE_GPU'] = 0
-        output['CE_core'] = 0
-        output['CE_memory'] = 0
-        output['n_treeMonths'] = 0
-        output['flying_context'] = 0
-        output['nkm_drivingUS'] = 0
-        output['nkm_drivingEU'] = 0
-        output['nkm_train'] = 0
-        output['energy_needed'] = 0
-        output['power_needed'] = 0
-        output['flying_text'] = None
-        output['text_CE'] = '... g CO2e'
-        output['appVersion'] = version
+#     ### If any of the required inputs is note ready: do not compute
+#     if notReady:
+#         output['coreType'] = None
+#         output['CPUmodel'] = None
+#         output['numberCPUs'] = None
+#         output['usageCPU'] = None
+#         output['usageCPUradio'] = None
+#         output['tdpCPU'] = None
+#         output['GPUmodel'] = None
+#         output['numberGPUs'] = None
+#         output['tdpGPU'] = None
+#         output['usageGPU'] = None
+#         output['usageGPUradio'] = None
+#         output['GPUpower'] = None
+#         output['memory'] = None
+#         output['runTime_hour'] = None
+#         output['runTime_min'] = None
+#         output['runTime'] = None
+#         output['platformType'] = None
+#         output['location'] = None
+#         output['carbonIntensity'] = None
+#         output['PUE'] = None
+#         output['PUEradio'] = None
+#         output['PSF'] = None
+#         output['PSFradio'] = None
+#         output['carbonEmissions'] = 0
+#         output['CE_CPU'] = 0
+#         output['CE_GPU'] = 0
+#         output['CE_core'] = 0
+#         output['CE_memory'] = 0
+#         output['n_treeMonths'] = 0
+#         output['flying_context'] = 0
+#         output['nkm_drivingUS'] = 0
+#         output['nkm_drivingEU'] = 0
+#         output['nkm_train'] = 0
+#         output['energy_needed'] = 0
+#         output['power_needed'] = 0
+#         output['flying_text'] = None
+#         output['text_CE'] = '... g CO2e'
+#         output['appVersion'] = version
 
-    #############################################
-    ### PRE-COMPUTATIONS: update variables used in the calcul based on inputs
+#     #############################################
+#     ### PRE-COMPUTATIONS: update variables used in the calcul based on inputs
 
-    else:
-        ### PUE
-        defaultPUE = data_dict.pueDefault_dict['Unknown']
-        # the input PUE is used only if the PUE box is shown AND the radio button is "Yes"
-        if (is_shown(PUEdivStyle)) & (PUEradio == 'Yes'):
-            PUE_used = PUE
+#     else:
+#         ### PUE
+#         defaultPUE = data_dict.pueDefault_dict['Unknown']
+#         # the input PUE is used only if the PUE box is shown AND the radio button is "Yes"
+#         if (is_shown(PUEdivStyle)) & (PUEradio == 'Yes'):
+#             PUE_used = PUE
         
-        ### PLATFORM ALONG WITH PUE
-        else:
-            if selected_platform == 'personalComputer':
-                PUE_used = 1
-            elif selected_platform == 'localServer':
-                PUE_used = defaultPUE
-            else:
-                # Cloud
-                if selected_provider == 'other':
-                    PUE_used = defaultPUE
-                else:
-                    # if we don't know the PUE of this specific data centre, or if we 
-                    # don't know the data centre, we use the provider's default
-                    server_data = data_dict.datacenters_dict_byName.get(server)
-                    if server_data is not None:
-                        if pd.isnull(server_data['PUE']):
-                            PUE_used = data_dict.pueDefault_dict[selected_provider]
-                        else:
-                            PUE_used = server_data['PUE']
-                    else:
-                        PUE_used = data_dict.pueDefault_dict[selected_provider]
+#         ### PLATFORM ALONG WITH PUE
+#         else:
+#             if selected_platform == 'personalComputer':
+#                 PUE_used = 1
+#             elif selected_platform == 'localServer':
+#                 PUE_used = defaultPUE
+#             else:
+#                 # Cloud
+#                 if selected_provider == 'other':
+#                     PUE_used = defaultPUE
+#                 else:
+#                     # if we don't know the PUE of this specific data centre, or if we 
+#                     # don't know the data centre, we use the provider's default
+#                     server_data = data_dict.datacenters_dict_byName.get(server)
+#                     if server_data is not None:
+#                         if pd.isnull(server_data['PUE']):
+#                             PUE_used = data_dict.pueDefault_dict[selected_provider]
+#                         else:
+#                             PUE_used = server_data['PUE']
+#                     else:
+#                         PUE_used = data_dict.pueDefault_dict[selected_provider]
 
-        ### CPUs
-        if coreType in ['CPU', 'Both']:
-            if is_shown(tdpCPUstyle):
-                # we asked the question about TDP
-                CPUpower = tdpCPU
-            else:
-                if CPUmodel == 'other':
-                    CPUpower = tdpCPU
-                else:
-                    CPUpower = data_dict.cores_dict['CPU'][CPUmodel]
-            if usageCPUradio == 'Yes':
-                usageCPU_used = usageCPU
-            else:
-                usageCPU_used = 1.
-            powerNeeded_CPU = PUE_used * n_CPUcores * CPUpower * usageCPU_used
-        else:
-            powerNeeded_CPU = 0
-            CPUpower = 0
-            usageCPU_used = 0
+#         ### CPUs
+#         if coreType in ['CPU', 'Both']:
+#             if is_shown(tdpCPUstyle):
+#                 # we asked the question about TDP
+#                 CPUpower = tdpCPU
+#             else:
+#                 if CPUmodel == 'other':
+#                     CPUpower = tdpCPU
+#                 else:
+#                     CPUpower = data_dict.cores_dict['CPU'][CPUmodel]
+#             if usageCPUradio == 'Yes':
+#                 usageCPU_used = usageCPU
+#             else:
+#                 usageCPU_used = 1.
+#             powerNeeded_CPU = PUE_used * n_CPUcores * CPUpower * usageCPU_used
+#         else:
+#             powerNeeded_CPU = 0
+#             CPUpower = 0
+#             usageCPU_used = 0
 
-        if coreType in ['GPU', 'Both']:
-            if is_shown(tdpGPUstyle):
-                GPUpower = tdpGPU
-            else:
-                if GPUmodel == 'other':
-                    GPUpower = tdpGPU
-                else:
-                    GPUpower = data_dict.cores_dict['GPU'][GPUmodel]
-            if usageGPUradio == 'Yes':
-                usageGPU_used = usageGPU
-            else:
-                usageGPU_used = 1.
-            powerNeeded_GPU = PUE_used * n_GPUs * GPUpower * usageGPU_used
-        else:
-            powerNeeded_GPU = 0
-            GPUpower = 0
-            usageGPU_used = 0
+#         if coreType in ['GPU', 'Both']:
+#             if is_shown(tdpGPUstyle):
+#                 GPUpower = tdpGPU
+#             else:
+#                 if GPUmodel == 'other':
+#                     GPUpower = tdpGPU
+#                 else:
+#                     GPUpower = data_dict.cores_dict['GPU'][GPUmodel]
+#             if usageGPUradio == 'Yes':
+#                 usageGPU_used = usageGPU
+#             else:
+#                 usageGPU_used = 1.
+#             powerNeeded_GPU = PUE_used * n_GPUs * GPUpower * usageGPU_used
+#         else:
+#             powerNeeded_GPU = 0
+#             GPUpower = 0
+#             usageGPU_used = 0
 
-        ### SERVER/LOCATION
-        carbonIntensity = data_dict.CI_dict_byLoc[locationVar]['carbonIntensity']
+#         ### SERVER/LOCATION
+#         carbonIntensity = data_dict.CI_dict_byLoc[locationVar]['carbonIntensity']
 
-        ### PSF
-        if PSFradio == 'Yes':
-            PSF_used = PSF
-        else:
-            PSF_used = 1
+#         ### PSF
+#         if PSFradio == 'Yes':
+#             PSF_used = PSF
+#         else:
+#             PSF_used = 1
 
-        #############################################
-        ### COMPUTATIONS: final outputs are computed
+#         #############################################
+#         ### COMPUTATIONS: final outputs are computed
 
-        # Power needed, in Watt
-        powerNeeded_core = powerNeeded_CPU + powerNeeded_GPU
-        powerNeeded_memory = PUE_used * (memory * data_dict.refValues_dict['memoryPower'])
-        powerNeeded = powerNeeded_core + powerNeeded_memory
+#         # Power needed, in Watt
+#         powerNeeded_core = powerNeeded_CPU + powerNeeded_GPU
+#         powerNeeded_memory = PUE_used * (memory * data_dict.refValues_dict['memoryPower'])
+#         powerNeeded = powerNeeded_core + powerNeeded_memory
 
-        # Energy needed, in kWh (so dividing by 1000 to convert to kW)
-        energyNeeded_CPU = runTime * powerNeeded_CPU * PSF_used / 1000
-        energyNeeded_GPU = runTime * powerNeeded_GPU * PSF_used / 1000
-        energyNeeded_core = runTime * powerNeeded_core * PSF_used / 1000
-        eneregyNeeded_memory = runTime * powerNeeded_memory * PSF_used / 1000
-        energyNeeded = runTime * powerNeeded * PSF_used / 1000
+#         # Energy needed, in kWh (so dividing by 1000 to convert to kW)
+#         energyNeeded_CPU = runTime * powerNeeded_CPU * PSF_used / 1000
+#         energyNeeded_GPU = runTime * powerNeeded_GPU * PSF_used / 1000
+#         energyNeeded_core = runTime * powerNeeded_core * PSF_used / 1000
+#         eneregyNeeded_memory = runTime * powerNeeded_memory * PSF_used / 1000
+#         energyNeeded = runTime * powerNeeded * PSF_used / 1000
 
-        # Carbon emissions: carbonIntensity is in g per kWh, so results in gCO2
-        CE_CPU = energyNeeded_CPU * carbonIntensity
-        CE_GPU = energyNeeded_GPU * carbonIntensity
-        CE_core = energyNeeded_core * carbonIntensity
-        CE_memory  = eneregyNeeded_memory * carbonIntensity
-        carbonEmissions = energyNeeded * carbonIntensity
+#         # Carbon emissions: carbonIntensity is in g per kWh, so results in gCO2
+#         CE_CPU = energyNeeded_CPU * carbonIntensity
+#         CE_GPU = energyNeeded_GPU * carbonIntensity
+#         CE_core = energyNeeded_core * carbonIntensity
+#         CE_memory  = eneregyNeeded_memory * carbonIntensity
+#         carbonEmissions = energyNeeded * carbonIntensity
 
-        # Storing all outputs to catch the app state and adapt textual content
-        output['coreType'] = coreType
-        output['CPUmodel'] = CPUmodel
-        output['numberCPUs'] = n_CPUcores
-        output['tdpCPU'] = CPUpower
-        output['usageCPUradio'] = usageCPUradio
-        output['usageCPU'] = usageCPU_used
-        output['GPUmodel'] = GPUmodel
-        output['numberGPUs'] = n_GPUs
-        output['tdpGPU'] = GPUpower
-        output['usageGPUradio'] = usageGPUradio
-        output['usageGPU'] = usageGPU_used
-        output['memory'] = memory
-        output['runTime_hour'] = actual_runTime_hours
-        output['runTime_min'] = actual_runTime_min
-        output['runTime'] = runTime
-        output['platformType'] = selected_platform
-        output['locationContinent'] = locationContinent
-        output['locationCountry'] = locationCountry
-        output['locationRegion'] = locationRegion
-        output['provider'] = selected_provider
-        output['serverContinent'] = serverContinent
-        output['server'] = server
-        output['location'] = locationVar
-        output['carbonIntensity'] = carbonIntensity
-        output['PUE'] = PUE_used
-        output['PUEradio'] = PUEradio
-        output['PSF'] = PSF_used
-        output['PSFradio'] = PSFradio
-        output['carbonEmissions'] = carbonEmissions
-        output['CE_CPU'] = CE_CPU
-        output['CE_GPU'] = CE_GPU
-        output['CE_core'] = CE_core
-        output['CE_memory'] = CE_memory
-        output['energy_needed'] = energyNeeded
-        output['power_needed'] = powerNeeded
-        output['appVersion'] = version
+#         # Storing all outputs to catch the app state and adapt textual content
+#         output['coreType'] = coreType
+#         output['CPUmodel'] = CPUmodel
+#         output['numberCPUs'] = n_CPUcores
+#         output['tdpCPU'] = CPUpower
+#         output['usageCPUradio'] = usageCPUradio
+#         output['usageCPU'] = usageCPU_used
+#         output['GPUmodel'] = GPUmodel
+#         output['numberGPUs'] = n_GPUs
+#         output['tdpGPU'] = GPUpower
+#         output['usageGPUradio'] = usageGPUradio
+#         output['usageGPU'] = usageGPU_used
+#         output['memory'] = memory
+#         output['runTime_hour'] = actual_runTime_hours
+#         output['runTime_min'] = actual_runTime_min
+#         output['runTime'] = runTime
+#         output['platformType'] = selected_platform
+#         output['locationContinent'] = locationContinent
+#         output['locationCountry'] = locationCountry
+#         output['locationRegion'] = locationRegion
+#         output['provider'] = selected_provider
+#         output['serverContinent'] = serverContinent
+#         output['server'] = server
+#         output['location'] = locationVar
+#         output['carbonIntensity'] = carbonIntensity
+#         output['PUE'] = PUE_used
+#         output['PUEradio'] = PUEradio
+#         output['PSF'] = PSF_used
+#         output['PSFradio'] = PSFradio
+#         output['carbonEmissions'] = carbonEmissions
+#         output['CE_CPU'] = CE_CPU
+#         output['CE_GPU'] = CE_GPU
+#         output['CE_core'] = CE_core
+#         output['CE_memory'] = CE_memory
+#         output['energy_needed'] = energyNeeded
+#         output['power_needed'] = powerNeeded
+#         output['appVersion'] = version
 
-        ### Context
-        output['n_treeMonths'] = carbonEmissions / data_dict.refValues_dict['treeYear'] * 12
-        output['nkm_drivingUS'] = carbonEmissions / data_dict.refValues_dict['passengerCar_US_perkm']
-        output['nkm_drivingEU'] = carbonEmissions / data_dict.refValues_dict['passengerCar_EU_perkm']
-        output['nkm_train'] = carbonEmissions / data_dict.refValues_dict['train_perkm']
+#         ### Context
+#         output['n_treeMonths'] = carbonEmissions / data_dict.refValues_dict['treeYear'] * 12
+#         output['nkm_drivingUS'] = carbonEmissions / data_dict.refValues_dict['passengerCar_US_perkm']
+#         output['nkm_drivingEU'] = carbonEmissions / data_dict.refValues_dict['passengerCar_EU_perkm']
+#         output['nkm_train'] = carbonEmissions / data_dict.refValues_dict['train_perkm']
 
-        ### Text plane trips
-        if carbonEmissions < 0.5 * data_dict.refValues_dict['flight_NY-SF']:
-            output['flying_context'] = carbonEmissions / data_dict.refValues_dict['flight_PAR-LON']
-            output['flying_text'] = "Paris-London"
-        elif carbonEmissions < 0.5 * data_dict.refValues_dict['flight_NYC-MEL']:
-            output['flying_context'] = carbonEmissions / data_dict.refValues_dict['flight_NY-SF']
-            output['flying_text'] = "NYC-San Francisco"
-        else:
-            output['flying_context'] = carbonEmissions / data_dict.refValues_dict['flight_NYC-MEL']
-            output['flying_text'] = "NYC-Melbourne"
+#         ### Text plane trips
+#         if carbonEmissions < 0.5 * data_dict.refValues_dict['flight_NY-SF']:
+#             output['flying_context'] = carbonEmissions / data_dict.refValues_dict['flight_PAR-LON']
+#             output['flying_text'] = "Paris-London"
+#         elif carbonEmissions < 0.5 * data_dict.refValues_dict['flight_NYC-MEL']:
+#             output['flying_context'] = carbonEmissions / data_dict.refValues_dict['flight_NY-SF']
+#             output['flying_text'] = "NYC-San Francisco"
+#         else:
+#             output['flying_context'] = carbonEmissions / data_dict.refValues_dict['flight_NYC-MEL']
+#             output['flying_text'] = "NYC-Melbourne"
 
-        ### Text carbon emissions
-        carbonEmissions_value = carbonEmissions  # in g CO2e
-        carbonEmissions_unit = "g"
-        if carbonEmissions_value >= 1e6:
-            carbonEmissions_value /= 1e6
-            carbonEmissions_unit = "T"
-        elif carbonEmissions_value >= 1e3:
-            carbonEmissions_value /= 1e3
-            carbonEmissions_unit = "kg"
-        elif carbonEmissions_value < 1:
-            carbonEmissions_value *= 1e3
-            carbonEmissions_unit = "mg"
-        if (carbonEmissions_value != 0)&((carbonEmissions_value >= 1e3)|(carbonEmissions_value < 1)):
-            output['text_CE'] = f"{carbonEmissions_value:,.2e} {carbonEmissions_unit} CO2e"
-        else:
-            output['text_CE'] = f"{carbonEmissions_value:,.2f} {carbonEmissions_unit} CO2e"
+#         ### Text carbon emissions
+#         carbonEmissions_value = carbonEmissions  # in g CO2e
+#         carbonEmissions_unit = "g"
+#         if carbonEmissions_value >= 1e6:
+#             carbonEmissions_value /= 1e6
+#             carbonEmissions_unit = "T"
+#         elif carbonEmissions_value >= 1e3:
+#             carbonEmissions_value /= 1e3
+#             carbonEmissions_unit = "kg"
+#         elif carbonEmissions_value < 1:
+#             carbonEmissions_value *= 1e3
+#             carbonEmissions_unit = "mg"
+#         if (carbonEmissions_value != 0)&((carbonEmissions_value >= 1e3)|(carbonEmissions_value < 1)):
+#             output['text_CE'] = f"{carbonEmissions_value:,.2e} {carbonEmissions_unit} CO2e"
+#         else:
+#             output['text_CE'] = f"{carbonEmissions_value:,.2f} {carbonEmissions_unit} CO2e"
 
-        ### Text energy
-        energyNeeded_value = energyNeeded  # in kWh
-        energyNeeded_unit = "kWh"
-        if energyNeeded_value >= 1e3:
-            energyNeeded_value /= 1e3
-            energyNeeded_unit = "MWh"
-        elif energyNeeded_value < 1:
-            energyNeeded_value *= 1e3
-            energyNeeded_unit = "Wh"
-        if (energyNeeded_value != 0) & ((energyNeeded_value >= 1e3) | (energyNeeded_value < 1)):
-            output['text_energyNeeded'] = f"{energyNeeded_value:,.2e} {energyNeeded_unit}"
-        else:
-            output['text_energyNeeded'] = f"{energyNeeded_value:,.2f} {energyNeeded_unit}"
+#         ### Text energy
+#         energyNeeded_value = energyNeeded  # in kWh
+#         energyNeeded_unit = "kWh"
+#         if energyNeeded_value >= 1e3:
+#             energyNeeded_value /= 1e3
+#             energyNeeded_unit = "MWh"
+#         elif energyNeeded_value < 1:
+#             energyNeeded_value *= 1e3
+#             energyNeeded_unit = "Wh"
+#         if (energyNeeded_value != 0) & ((energyNeeded_value >= 1e3) | (energyNeeded_value < 1)):
+#             output['text_energyNeeded'] = f"{energyNeeded_value:,.2e} {energyNeeded_unit}"
+#         else:
+#             output['text_energyNeeded'] = f"{energyNeeded_value:,.2f} {energyNeeded_unit}"
 
-        ### Text tree-months
-        treeTime_value = output['n_treeMonths']  # in tree-months
-        treeTime_unit = "tree-months"
-        if treeTime_value >= 24:
-            treeTime_value /= 12
-            treeTime_unit = "tree-years"
-        if (treeTime_value != 0) & ((treeTime_value >= 1e3) | (treeTime_value < 0.1)):
-            output['text_treeYear'] = f"{treeTime_value:,.2e} {treeTime_unit}"
-        else:
-            output['text_treeYear'] = f"{treeTime_value:,.2f} {treeTime_unit}"
+#         ### Text tree-months
+#         treeTime_value = output['n_treeMonths']  # in tree-months
+#         treeTime_unit = "tree-months"
+#         if treeTime_value >= 24:
+#             treeTime_value /= 12
+#             treeTime_unit = "tree-years"
+#         if (treeTime_value != 0) & ((treeTime_value >= 1e3) | (treeTime_value < 0.1)):
+#             output['text_treeYear'] = f"{treeTime_value:,.2e} {treeTime_unit}"
+#         else:
+#             output['text_treeYear'] = f"{treeTime_value:,.2f} {treeTime_unit}"
 
-    return output
+#     return output
 
 
-## UPDATE TOP TEXT
-##################
+# ## UPDATE TOP TEXT
+# ##################
 
 @app.callback(
     [
@@ -1356,7 +1362,7 @@ def aggregate_input_values(data, coreType, n_CPUcores, CPUmodel, tdpCPUstyle, td
         Output("driving_text", "children"),
         Output("flying_text", "children"),
     ],
-    [Input("aggregate_data", "data")],
+    [Input(form_ids.aggregate_data(ID_MAIN_FORM), "data")],
 )
 def update_text(data):
     text_CE = data.get('text_CE')
@@ -1382,7 +1388,7 @@ def update_text(data):
 
 @app.callback(
     Output("flying_label", "children"),
-    [Input("aggregate_data", "data")],
+    Input(form_ids.aggregate_data(ID_MAIN_FORM), "data"),
 )
 def update_text(data):
     if (data['flying_context'] >= 1)|(data['flying_context'] == 0):
@@ -1427,7 +1433,7 @@ def flush_input_csv_content(n):
 @app.callback(
     Output("aggregate-data-csv", "data"),
     Input("btn-download_csv", "n_clicks"),
-    State("aggregate_data", "data"),
+    State(form_ids.aggregate_data(ID_MAIN_FORM), "data"),
     prevent_initial_call=True,
 )
 def export_as_csv(_, aggregate_data):
@@ -1440,12 +1446,12 @@ def export_as_csv(_, aggregate_data):
     return dcc.send_data_frame(to_export.to_csv, f"GreenAlgorithms_results_{now}.csv", index=False, sep=';')
 
 
-## OUTPU GRAPHICS
+## OUTPUT GRAPHICS
 #################
 
 @app.callback(
     Output("pie_graph", "figure"),
-    [Input("aggregate_data", "data")],
+    Input(form_ids.aggregate_data(ID_MAIN_FORM), "data"),
 )
 def create_pie_graph(aggData):
     return create_cores_memory_pie_graphic(aggData)
@@ -1455,7 +1461,7 @@ def create_pie_graph(aggData):
 @app.callback(
     Output("barPlotComparison", "figure"),
     [
-        Input("aggregate_data", "data"),
+        Input(form_ids.aggregate_data(ID_MAIN_FORM), "data"),
         Input('versioned_data','data')
     ],
 )
@@ -1469,7 +1475,7 @@ def create_bar_chart(aggData, data):
 @app.callback(
     Output("barPlotComparison_cores", "figure"),
     [
-        Input("aggregate_data", "data"),
+        Input(form_ids.aggregate_data(ID_MAIN_FORM), "data"),
         Input('versioned_data','data')
     ],
 )
@@ -1488,7 +1494,7 @@ def create_bar_chart_cores(aggData, data):
 @app.callback(
     Output('report_markdown', 'children'),
     [
-        Input("aggregate_data", "data"),
+        Input(form_ids.aggregate_data(ID_MAIN_FORM), "data"),
         Input('versioned_data','data')
     ],
 )
