+++ pages/html_layout.py
@@ -2,12 +2,12 @@ import os
 import dash
 
 from dash import html, dcc
-from utils.handle_inputs import get_available_versions, CURRENT_VERSION
-from utils.utils import YES_NO_OPTIONS
+from utils.handle_inputs import get_available_versions
 from utils.graphics import BLANK_FIGURE
 
+from all_in_one_components.form.green_algo_form_AIO import GreenAlgoFormAIO, ID_MAIN_FORM
+
 import dash_bootstrap_components as dbc
-import plotly.graph_objects as go
 
 dash.register_page(__name__, path='/', title='Green Algorithms')
 
@@ -63,7 +63,7 @@ def layout(**query_strings):
                         " discussing different options for carbon footprint estimation."
                     ]),
 
-            #         # TODO add something else there? GA4HPC?
+                    # TODO add something else there? GA4HPC?
 
                     html.Div(
                         [
@@ -86,619 +86,7 @@ def layout(**query_strings):
 
             #### INPUT FORM ####
 
-            html.Form(
-                [
-                    html.H2(
-                        "Details about your algorithm"
-                    ),
-                    html.Center(
-                        html.P(["To understand how each parameter impacts your carbon footprint, "
-                                "check out the formula below and the ",
-                                html.A("methods article",
-                                       href='https://onlinelibrary.wiley.com/doi/10.1002/advs.202100707',
-                                       target='_blank')
-                                ]),
-                    ),
-
-                    ## RUN TIME
-                    html.Div(
-                        [
-                            html.Label("Runtime (HH:MM)"),
-
-                            html.Div(
-                                [
-                                    dcc.Input(
-                                        type='number',
-                                        id="runTime_hour_input",
-                                        min=0,
-                                    ),
-
-                                    dcc.Input(
-                                        type='number',
-                                        id="runTime_min_input",
-                                        min=0,
-                                        max=59,
-                                    )
-                                ],
-                                className="box-runtime box-fields"
-                            ),
-                        ],
-                        className='form-row short-input'
-                    ),
-
-                    html.Div(
-                        [
-                            html.Hr(),
-                        ],
-                        className='Hr_div'
-                    ),
-
-                    ## TYPE OF CORES
-                    html.Div(
-                        [
-                            html.Label("Type of cores"),
-
-                            html.Div(
-                                [
-                                    dcc.Dropdown(
-                                        id="coreType_dropdown",
-                                        clearable=False,
-                                    ),
-                                ],
-                                className="box-fields"
-                            ),
-
-                            html.Div(
-                                [
-                                    html.Div('i', className='tooltip-icon'),
-                                    html.P(
-                                        "Select the type of hardware used.",
-                                        className='tooltip-text'
-                                    ),
-                                ],
-                                className='tooltip',
-                            ),
-                        ],
-                        className='form-row short-input'
-                    ),
-
-                    ## CPUs
-                    html.Div(
-                        [
-                            html.H3(
-                                "CPUs",
-                                id='title_CPU'
-                            ),
-
-                            html.Div(
-                                [
-                                    html.Label("Number of cores"),
-
-                                    dcc.Input(
-                                        type='number',
-                                        id="numberCPUs_input",
-                                        min=0,
-                                    ),
-
-                                    html.Div(
-                                        [
-                                            html.Div('i', className='tooltip-icon'),
-                                            html.P(
-                                                "A single CPU contains several cores.",
-                                                className='tooltip-text'
-                                            ),
-                                        ],
-                                        className='tooltip',
-                                    ),
-                                ],
-                                className='form-row short-input'
-                            ),
-
-                            html.Div(
-                                [
-                                    html.Label("Model"),
-
-                                    html.Div(
-                                        [
-                                            dcc.Dropdown(
-                                                id="CPUmodel_dropdown",
-                                                className='bottom-dropdown',
-                                                clearable=False,
-                                            ),
-                                        ],
-                                        className="box-fields"
-                                    ),
-
-                                    html.Div(
-                                        [
-                                            html.Div('i', className='tooltip-icon'),
-                                            html.P(
-                                                "To fill-in a custom core power usage (TDP), please select 'other'.",
-                                                className='tooltip-text'
-                                            ),
-                                        ],
-                                        className='tooltip',
-                                    ),
-                                    
-                                ],
-                                className='form-row short-input'
-                            ),
-
-                            # CPU TDP
-                            html.Div(
-                                [
-                                    html.Label(
-                                        'What is the Thermal Design Power (TDP) value per core of your CPU? '
-                                        'This can easily be found online (usually 10-15W per core)'),
-
-                                    dcc.Input(
-                                        type='number',
-                                        id="tdpCPU_input",
-                                        min=0,
-                                    )
-                                ],
-                                className='form-row',
-                                id='tdpCPU_div',
-                                style=dict(display='none')
-                            ),
-                        ],
-                        className="group-of-rows",
-                        id="CPU_div"
-                    ),
-
-                    ## GPUs
-                    html.Div(
-                        [
-                            html.H3(
-                                "GPUs",
-                                id='title_GPU'
-                            ),
-
-                            html.Div(
-                                [
-                                    html.Label("Number of GPUs"),
-
-                                    dcc.Input(
-                                        type='number',
-                                        id="numberGPUs_input",
-                                        min=0,
-                                    ),
-
-                                    html.Div(
-                                        [
-                                            html.Div('i', className='tooltip-icon'),
-                                            html.P(
-                                                "Here, we ask for the number of GPUs (not the cores).",
-                                                className='tooltip-text'
-                                            ),
-                                        ],
-                                        className='tooltip',
-                                    ),
-                                ],
-                                className='form-row short-input'
-                            ),
-
-                            html.Div(
-                                [
-                                    html.Label("Model"),
-
-                                    html.Div(
-                                        [
-                                            dcc.Dropdown(
-                                                id="GPUmodel_dropdown",
-                                                className='bottom-dropdown',
-                                                clearable=False,
-                                            ),
-                                        ],
-                                        className="box-fields"
-                                    ),
-
-                                    html.Div(
-                                        [
-                                            html.Div('i', className='tooltip-icon'),
-                                            html.P(
-                                                "To fill-in a custom TDP, please select 'other'.",
-                                                className='tooltip-text'
-                                            ),
-                                        ],
-                                        className='tooltip',
-                                    ),
-                                ],
-                                className='form-row short-input'
-                            ),
-
-                            # GPU TDP
-                            html.Div(
-                                [
-                                    html.Label(
-                                        'What is the Thermal Design Power (TDP) value per core of your GPU? '
-                                        'This can easily be found online (usually around 200W)'),
-
-                                    dcc.Input(
-                                        type='number',
-                                        id="tdpGPU_input",
-                                        min=0,
-                                    )
-                                ],
-                                className='form-row',
-                                id='tdpGPU_div',
-                                style=dict(display='none')
-                            ),
-                        ],
-                        className="group-of-rows",
-                        id="GPU_div"
-                    ),
-
-                    html.Div(
-                        [
-                            html.Hr(),
-                        ],
-                        className='Hr_div'
-                    ),
-
-                    ## MEMORY
-                    html.Div(
-                        [
-                            html.Label("Memory available (in GB)"),
-
-                            dcc.Input(
-                                type='number',
-                                id="memory_input",
-                                min=0,
-                            ),
-
-                            html.Div(
-                                [
-                                    html.Div('i', className='tooltip-icon'),
-                                    html.P(
-                                        "We refer to the ‘allocated memory’, not the memory actually used by the program.",
-                                        className='tooltip-text'
-                                    ),
-                                ],
-                                className='tooltip',
-                            ),
-                        ],
-                        className='form-row short-input',
-                        id='div_memory'
-                    ),
-
-                    html.Div(
-                        [
-                            html.Hr(),
-                        ],
-                        className='Hr_div'
-                    ),
-
-                    ## SELECT COMPUTING PLATFORM
-                    html.Div(
-                        [
-                            html.Label("Select the platform used for the computations"),
-
-                            html.Div(
-                                [
-                                    dcc.Dropdown(
-                                        id="platformType_dropdown",
-                                        clearable=False,
-                                    ),
-                                ],
-                                className='box-fields',
-                            ),
-
-                            html.Div(
-                                [
-                                    html.Div('i', className='tooltip-icon'),
-                                    html.P(
-                                        "This field enables us to retrieve the PUE and energy mix associated with your computations.",
-                                        className='tooltip-text'
-                                    ),
-                                ],
-                                className='tooltip',
-                            ),    
-
-                            ## SELECT PROVIDER, FOR CLOUD COMPUTING ONLY
-                            html.Div(
-                                [
-                                    dcc.Dropdown(
-                                        id="provider_dropdown",
-                                        clearable=False,
-                                        className='bottom-dropdown',
-                                    ),
-                                ],
-                                className="box-fields",
-                                id="provider_dropdown_div"
-                            ),
-                        ],
-                        className='form-row'
-                    ),
-
-
-                    ## SERVER (for cloud computing)
-                    html.Div(
-                        [
-                            html.Label("Select server"),
-
-                            html.Div(
-                                [
-                                    dcc.Dropdown(
-                                        id="server_continent_dropdown",
-                                        clearable=False,
-                                    ),
-
-                                    dcc.Dropdown(
-                                        id="server_dropdown",
-                                        clearable=False,
-                                        className='bottom-dropdown',
-                                    ),
-                                ],
-                                className="box-fields"
-                            )
-                        ],
-                        id='server_div',
-                        className='form-row',
-                        style={'display': 'none'}
-                    ),
-
-                    ## LOCATION
-                    html.Div(
-                        [
-                            html.Label("Select location"),
-
-                            html.Div(
-                                [
-                                    dcc.Dropdown(
-                                        id="location_continent_dropdown",
-                                        clearable=False,
-                                    ),
-                                ],
-                                className="box-fields",
-                            ),
-
-                            html.Div(
-                                [
-                                    html.Div('i', className='tooltip-icon'),
-                                    html.P(
-                                        "In particular, the location is used to retrieve the energy mix.",
-                                        className='tooltip-text'
-                                    ),
-                                ],
-                                className='tooltip',
-                            ),   
-
-                            html.Div(
-                                [
-                                    dcc.Dropdown(
-                                        id="location_country_dropdown",
-                                        className='bottom-dropdown',
-                                        clearable=False,
-                                    ),
-                                ],
-                                className="box-fields",
-                                id="location_country_dropdown_div"
-                            ),
-
-                            html.Div(
-                                [
-                                    dcc.Dropdown(
-                                        id="location_region_dropdown",
-                                        className='bottom-dropdown',
-                                        clearable=False,
-                                    ),
-                                ],
-                                className="box-fields",
-                                id="location_region_dropdown_div"
-                            ),
-                        ],
-                        id='location_div',
-                        className='form-row',
-                        style={'display': 'flex'}
-                    ),
-
-                    ## CORE USAGE (CPU and GPU)
-                    html.Div(
-                        [
-                            html.Label("Do you know the real usage factor of your CPU?"),
-                            
-                            html.Div(
-                                [
-                                    dcc.RadioItems(
-                                        id='usageCPU_radio',
-                                        options=YES_NO_OPTIONS,
-                                        className="radio-input",
-                                    ),
-                                    dcc.Input(
-                                        min=0,
-                                        max=1,
-                                        # step=0.1,
-                                        type='number',
-                                        id="usageCPU_input",
-                                        style=dict(display='none'),
-                                    )
-                                ],
-                                className='radio-and-field'
-                            ),
-
-                            html.Div(
-                                [
-                                    html.Div('i', className='tooltip-icon'),
-                                    html.P(
-                                        "Between 0 and 1. Should correspond to a temporal factor usage, accessible from log files for instance.",
-                                        className='tooltip-text'
-                                    ),
-                                ],
-                                className='tooltip',
-                            ),
-                        ],
-                        className='form-row radio-row',
-                        id='usageCPU_div'
-                    ),
-
-                    html.Div(
-                        [
-                            html.Label("Do you know the real usage factor of your GPU?"),
-                            html.Div(
-                                [
-                                    dcc.RadioItems(
-                                        id='usageGPU_radio',
-                                        options=YES_NO_OPTIONS,
-                                        className="radio-input",
-                                    ),
-
-                                    dcc.Input(
-                                        min=0,
-                                        max=1,
-                                        # step=0.1,
-                                        type='number',
-                                        id="usageGPU_input",
-                                        style=dict(display='none'),
-                                    ),
-                                ],
-                                className='radio-and-field'
-                            ),
-
-                            html.Div(
-                                [
-                                    html.Div('i', className='tooltip-icon'),
-                                    html.P(
-                                        "Between 0 and 1. In most cases, should be 1.",
-                                        className='tooltip-text'
-                                    ),
-                                ],
-                                className='tooltip',
-                            ),                            
-                        ],
-                        className='form-row radio-row',
-                        id='usageGPU_div'
-                    ),
-
-                    ## PUE
-                    html.Div(
-                        [
-                            html.Label("Do you know the Power Usage Efficiency (PUE) of your local data centre?"),
-                            html.Div(
-                                [
-                                    dcc.RadioItems(
-                                        id='pue_radio',
-                                        options=YES_NO_OPTIONS,
-                                        className='radio-input',
-                                    ),
-
-                                    dcc.Input(
-                                        min=1,
-                                        type='number',
-                                        id="PUE_input",
-                                        style=dict(display='none'),
-                                    ),
-                                ],
-                                className='radio-and-field'
-                            ),
-
-                            html.Div(
-                                [
-                                    html.Div('i', className='tooltip-icon'),
-                                    html.P(
-                                        "The PUE accounts for the extra energy consumption of the data centre, including cooling (1.67 by default).",
-                                        className='tooltip-text'
-                                    ),
-                                ],
-                                className='tooltip',
-                            ),
-
-                        ],
-                        className='form-row radio-row',
-                        id='PUEquestion_div',
-                        style=dict(display='none'),
-                    ),
-
-                    ## PSF
-                    html.Div(
-                        [
-                            html.Label("Do you want to use a Pragmatic Scaling Factor (PSF)?"),
-                            html.Div(
-                                [
-                                    dcc.RadioItems(
-                                        id='PSF_radio',
-                                        options=YES_NO_OPTIONS,
-                                        className="radio-input",
-                                    ),
-
-                                    dcc.Input(
-                                        min=1,
-                                        type='number',
-                                        id="PSF_input",
-                                        style=dict(display='none'),
-                                    ),
-                                ],
-                                className='radio-and-field',
-                            ),
-
-                            html.Div(
-                                [
-                                    html.Div('i', className='tooltip-icon'),
-                                    html.P(
-                                        "The PSF refers to the number of repetions of the computation.",
-                                        className='tooltip-text'
-                                    ),
-                                ],
-                                className='tooltip',
-                            ),
-
-                        ],
-                        className='form-row radio-row',
-                    ),
-
-                    dcc.ConfirmDialog(
-                        id='confirm_reset',
-                        message='This will reset all the values you have entered so far!\n' \
-                        'Are you sure you want to continue?',
-                    ),
-
-                    html.Div(
-                        [
-                            html.Div(
-                                [
-                                    html.P("Reset", id='reset_link'),
-                                ],
-                                className='reset-button'
-                            ),
-                            html.Div(
-                                [
-                                    html.P("Change app version", id='oldVersion_link'),
-                                ],
-                                className='reset-button'
-                            ),
-                        ],
-                        className='two-buttons'
-                    ),
-
-                    html.Div(
-                        [
-                            html.Label("App version"),
-
-                            html.Div(
-                                [
-                                    dcc.Dropdown(
-                                        id="appVersions_dropdown",
-                                        options=appVersions_options,
-                                        className='bottom-dropdown',
-                                        clearable=False,
-                                    ),
-                                ],
-                                className="box-fields"
-                            )
-                        ],
-                        className='form-row short-input',
-                        id='oldVersions_div',
-                        style=dict(display='none'),
-                    ),
-
-                    html.P(
-                        id="placeholder",
-                        style={"display": "none"}
-                    )
-
-                ],
-                className='container input-form'
-            ),
+            GreenAlgoFormAIO(ID_MAIN_FORM),
 
             #### FIRST OUTPUTS ####
 
@@ -905,7 +293,7 @@ def layout(**query_strings):
 
                     dcc.Interval(
                         id='csv-input-timer',
-                        interval=2000, 
+                        interval=10000, 
                         # in milliseconds, should not be lower than 1000
                         # otherwise the update of the upload csv content is done too soon
                         # and there is not consistency between the state of the form and
